+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add configuration to show the shelf title.
+
+### Changed
+- Add `product-summary` as a block in `interfaces.json`.
 
 ## [1.6.0] - 2019-02-15
 ### Added